requirements.txt
absl-py==2.0.0
accelerate==0.30.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
alias-free-torch==0.0.6
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.3.0
argbind==0.3.9
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audiocraft==1.2.0
audioread==3.0.1
av==11.0.0
backcall==0.2.0
beartype==0.18.5
bitarray==2.9.2
blis==0.7.11
boto3==1.29.6
botocore==1.32.6
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
clean-fid==0.1.35
click==8.1.7
clip-anytorch==2.6.0
cloudpathlib==0.16.0
cloudpickle==3.0.0
cn2an==0.5.22
colorama==0.4.6
colorlog==6.8.2
confection==0.1.4
contourpy==1.1.1
crcmod==1.7
cryptography==43.0.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dataclasses==0.6
datasets==2.18.0
dctorch==0.1.2
decorator==5.1.1
decord==0.6.0
deepspeed==0.14.0
demucs==4.0.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
diffusers==0.27.2
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
dora_search==0.1.12
einops==0.8.0
einops-exts==0.0.4
einx==0.3.0
ema-pytorch==0.5.1
encodec==0.1.1
exceptiongroup==1.2.0
executing==2.2.0
expecttest==0.1.6
fairseq==0.12.2
fastapi==0.110.3
ffmpy==0.3.2
filelock==3.13.1
fire==0.7.0
flashy==0.0.2
flatten-dict==0.4.2
fonttools==4.49.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2023.10.0
ftfy==6.1.3
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.4
google-auth-oauthlib==1.0.0
gradio==4.26.0
gradio_client==0.15.1
grpcio==1.59.3
h11==0.14.0
h5py==3.11.0
hf-xet==1.1.2
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.25.2
hydra-colorlog==1.2.0
hydra-core==1.0.7
hypothesis==6.90.0
idna==3.4
imageio==2.35.1
importlib-metadata==6.8.0
importlib_resources==6.1.3
inflect==7.0.0
ipython==8.12.3
jedi==0.19.2
jieba-fast==0.53
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json5==0.9.25
jsonlines==4.0.0
jsonmerge==1.9.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
julius==0.2.7
k-diffusion==0.1.1.post1
kaldiio==2.18.0
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.9
lameenc==1.7.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.3
librosa==0.9.1
lightning==2.2.1
lightning-utilities==0.10.1
lion-pytorch==0.2.2
llvmlite==0.41.1
local-attention==1.9.14
loguru==0.7.2
lxml==5.2.2
marisa-trie==1.1.1
Markdown==3.5.1
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.16.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1.1
nltk==3.8.1
nnAudio==0.3.3
num2words==0.5.13
numba==0.58.1
numpy==1.23.5
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.7.0.84
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.19.3
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
omegaconf==2.2.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openunmix==1.2.1
orjson==3.10.3
oss2==2.18.6
packaging==23.2
pandas==2.0.3
parso==0.8.4
peft==0.10.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
pooch==1.8.1
portalocker==2.10.1
preshed==3.0.9
proces==0.1.7
prodict==0.8.18
progressbar==2.5
prompt_toolkit==3.0.51
protobuf==3.19.6
psutil==5.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pynvml==11.5.0
pyparsing==3.1.2
pypinyin==0.51.0
pyre-extensions==0.0.29
pyreaper==0.0.10
PySoundFile==0.9.0.post1
pystoi==0.4.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch-lightning==2.2.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
randomname==0.2.1
referencing==0.35.1
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
resampy==0.4.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
ruff==0.4.4
s3transfer==0.7.0
sacrebleu==2.4.2
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.10.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sox==1.4.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
submitit==1.5.1
sympy==1.12
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
thinc==8.2.3
threadpoolctl==3.3.0
tifffile==2023.7.10
timm==0.9.11
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.0+cu118
torch-stoi==0.2.3
torchaudio==2.2.0+cu118
torchdata==0.7.1
torchdiffeq==0.2.5
torchlibrosa==0.1.0
torchmetrics==1.3.1
torchsde==0.2.6
torchtext==0.17.0
torchvision==0.17.0+cu118
tqdm==4.66.4
traitlets==5.14.3
trampoline==0.1.2
transformers==4.37.2
treetable==0.2.5
triton==2.2.0
typeguard==2.13.0
typer==0.9.4
types-dataclasses==0.6.6
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.8
urllib3==1.26.18
uvicorn==0.29.0
vector_quantize_pytorch==1.14.8
wandb==0.17.4
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webdataset==0.2.86
websockets==11.0.3
Werkzeug==3.0.1
wget==3.2
wordsegment==1.3.1
x-clip==0.14.4
x-transformers==2.0.0
xformers==0.0.24+cu118
yarl==1.9.4
zipp==3.17.0
